variable rename
Decided to give CLion a shot after Doom-Emacs decided it was going to do some stupid "Let Me Help You\!" types of bullshit. It kept flagging these two variables as being badly named and needing a change. As I'm not looking at the docs for this (instead I have Discord on that screen) while making the change, I think I fucked it up, but...