vector instructions largely decoded
There are some features where there are flags and adjunct registers in the same 6-bit space according to the docs. I'm not sure WTF is going on with them or how to properly represent them in the decode, so I've left them unimplemented for now. Vector48 needs a major rewrite at this point and Vector80 can do with a refactor to move out shared code to helpers.