More major refactoring work
This will not work fully as some of the debugging that was left in is
managing to trip a corner-case that will be addressed when SCALAR48 gets
its refactor in the next day or two.

At this time SCALAR16, SCALAR32, VECTOR48 and VECTOR80 are largely clean
code with none of the really confusing logic that pervaded the code
until recently. I have introduced a set of macros that rely on a pair of
other macros being defined that take the place of a lot of boilerplate
and clean up a lot of screen cruft.

In the future similar macros are likely to be introduced for bit-masks,
bit-shifts and magic-number situations as well as, possibly, a few of
the more common boilerplate constructs.