Last bit of refactoring
SCALAR48 now refactored properly. Oddly there seems to be a bug where a
properly decoded instruction is returning with the wrong size - I am
going to have to (possibly) take a look at what the various space-saving
macros are doing, but...

Otherwise things are working great and the last of the places where NULL
was returned have been replaced. Next I'll clean up the code in
'src/driver.cpp' to try and make it a lot more clear what is happening
and then its time for all the various QOL improvements that will be
needed.